rename files for consistency like-this
