+++ examples/next-openai/chat-components/providers.tsx
@@ -5,7 +5,7 @@ import { ThemeProviderProps } from 'next-themes/dist/types'
 
 import { TooltipProvider } from '@/chat-components/ui/tooltip'
 import { SidebarProvider } from '@/app/sidebar/sidebar-context'
-import { CopilotProvider } from '@/app/CopilotContext'
+import { CopilotProvider } from '@/app/copilot-context'
 import { ThemeProvider } from 'next-themes'
 
 export function Providers({ children, ...props }: ThemeProviderProps) {
